Health Synchronisation
There were some animation offsets so one client would detect a hit on an enemmy but another client would not. Using OnPhotonSerializeView from Photon, I made so that the health is compared between the same objects but different clients and fix the health of each once there was a difference detected. I had to change the Synchronization from "Unreliable On Change" to "Unreliable" because UOC checks only when the OTHER clients' values change, not when MINE changes. "Unreliable" sends the values and compares them not matter what, even if no changes were made.

Unreliable vs Delta Compress:
https://forum.unity.com/threads/reliable-delta-compressed-vs-unreliable.167050/